Register the ProfileIOGroup and EpochIOGroup in special accessor
