+++ Makefile.am
@@ -493,6 +493,8 @@ libgeopmpolicy_la_SOURCES = src/Accumulator.cpp \
                             src/PowerGovernorImp.hpp \
                             src/PowerGovernorAgent.cpp \
                             src/PowerGovernorAgent.hpp \
+                            src/PlatformIOProf.cpp \
+                            src/PlatformIOProf.hpp \
                             src/ProfileIOGroup.cpp \
                             src/ProfileIOGroup.hpp \
                             src/ProfileSampler.cpp \
@@ -527,8 +529,6 @@ libgeopmpolicy_la_SOURCES = src/Accumulator.cpp \
                             src/geopm_endpoint.h \
                             src/geopm_imbalancer.h \
                             src/geopm_policystore.h \
-                            src/profile_io_prof.cpp \
-                            src/profile_io_prof.hpp \
                             src/record.cpp \
                             src/record.hpp \
                             # end
