+++ src/PlatformIOProf.hpp
@@ -30,51 +30,24 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "platform_io_prof.hpp"
+#ifndef PLATFORM_IO_PROF_HPP_INCLUDE
+#define PLATFORM_IO_PROF_HPP_INCLUDE
 
-#include "geopm/PlatformIO.hpp"
-#include "ProfileIOGroup.hpp"
-#include "EpochIOGrup.hpp"
+#include <string>
 
 namespace geopm
 {
-    static PlatformIO &PlatformIOProf::platform_io(void)
-    {
-                static PlatformIOProf instance;
-        return instance.m_platform_io;
-    }
-    PlatformIOProf::PlatformIOProf()
-        : m_platform_io(platform_io())
-    {
-        try {
-            m_platform_io.register_iogroup(
-                ProfileIOGroup::make_plugin());
-        }
-        catch (const geopm::Exception &ex) {
-            print_load_warning("ProfileIOGroup", ex.what());
-        }
-        try {
-            m_platform_io.register_iogroup(
-                EpochIOGroup::make_plugin());
-        }
-        catch (const geopm::Exception &ex) {
-            print_load_warning("EpochIOGroup", ex.what());
-        }
-
-    }
-    void PlatformIOProf::print_load_warning(const std::string io_group_name,
-                                            const std::string what) const
-    {
-#ifdef GEOPM_DEBUG
-        std::cerr << "Warning: <geopm> Failed to load " << io_group_name
-                  << " IOGroup.  "
-                  << "GEOPM may not work properly unless an alternate "
-                  << "IOGroup plugin is loaded to provide signals/controls "
-                  << "required."
-                  << std::endl;
-        std::cerr << "The error was: " << what << std::endl;
-#endif
-    }
+    class PlatformIO;
+    class PlatformIOProf {
+        public:
+            static PlatformIO &platform_io(void);
+            virtual ~PlatformIOProf() = default;
+        private:
+            PlatformIOProf();
+            void print_load_warning(const std::string &io_group_name,
+                                    const std::string &what) const;
+            PlatformIO &m_platform_io;
+    };
 }
 
 #endif
