+++ src/Tracer.cpp
@@ -47,6 +47,7 @@
 #include "geopm/Exception.hpp"
 #include "geopm/Helper.hpp"
 #include "Environment.hpp"
+#include "PlatformIOProf.hpp"
 #include "geopm_hash.h"
 #include "geopm_version.h"
 #include "geopm.h"
@@ -57,8 +58,8 @@ namespace geopm
 {
     TracerImp::TracerImp(const std::string &start_time)
         : TracerImp(start_time, environment().trace(), hostname(),
-                    environment().do_trace(), platform_io(), platform_topo(),
-                    environment().trace_signals())
+                    environment().do_trace(), PlatformIOProf::platform_io(),
+                    platform_topo(), environment().trace_signals())
     {
 
     }
